requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
gunicorn==23.0.0
google-cloud-firestore==2.18.0
google-cloud-aiplatform==1.63.0
google-generativeai==0.7.2
python-dotenv==1.0.1
